backend/requirements.txt
requests
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
channels==3.0.5


